Temporary Message Initiation
In this commit, a bit of information has been added to the different
sections of the homepage. The alignment is wildly off after making one
emendation, and it is my belief that the issue originates in some
parallax scrolling class here (not to place blame...I probably am just
writing something conflicting with properties).